Release version 2.2.4
This fixes a small defect in the Open Resource command palette item,
which has an argument that indicates whether or not unknown overrides
(files which appear in an unpacked package but not in the
sublime-package file) should appear in the list.

The argument was not being used and always used a value of False,
meaning that when browsing for an override to open, unknown overrides
were not open-able.