Include meta scopes for package specifiers
In this context, a "package specifier" is the text `[SIU]` that appears
in override and diff reports, and the `[S]`, `[I]` and `[U]` values in
the package report that tell you what kind of package it is.

All syntaxes now apply meta scopes to these things, and the tests have
been augmented as well to verify them. This will allow us to create
more hover popups.