Set up list of pkgs with predefined plugin hosts
Previous to this, the code in the PackageInfo class would assume that
the User package runs in the 3.8 host if it's a ST4 build, and all
other packages run in either the legacy 3.3 host or, if they
have .plugin-version file, in that host.

The Default package is special in that it "straddles" the plugin host
and is available in both of them in a way that other plugins aren't.
This[ this alters the code so that we can say that the Default package
will run in multiple hosts where available in the package popup, which
is more accurate.