Provide initial support for info hover popups
Include the ability to provide hover popups that provide more
information on specific topics, as outlined by the text being hovered
over.

This is implemented for the package type specifiers in reports
(including the specific ones in a Package report) as well as for the
markers for the different types of overrides (expired, unknown,
complete, expired complete).

When hovering over some items, links are available in the popup that
will execute commands, including creating a diff report or a new
override, both in the package that the item hovered over is associated
with.