Release version 2.2.5
This release brings a couple of minor changes, most notably in the hover
popups that we provide on reports. The Default package is now more
visibly documented to be available in the 3.3 and 3.8 hosts, and there
is now hover help for some of the report specifiers, such as [X] and
[S] and the shorter form [SIU] varieties.

This build is the last build that is going to be made available for
Sublime Text 3 users.