Release version 2.2.6
There are no functional changes in this release; the only change here is
that the URL for the documentation has been updated in a few places
because the domain and hosting that is being used for the documentation
has been updated.

As such, this release is also being pushed for ST3 users, just so that
people using older versions can still reach the documentation site
without getting a 404 error.