Upgrade OverrideAudit to the latest plugin host
This is a small change to make the package run from within the 3.8
plugin host instead of the legacy 3.3 plugin host.

There should be no functional differences since the package does not use
any dependencies of any kind and packs everything in that it needs.

No material code changes have been made as of yet other than the control
file, so there are various things that could be updated as we go
forward, such as changes to format strings rather than calls to format()
for example.