ZipFile.open() no longer supports rU directly
For the purposes of making diffing files easier, we want to open all
resources as universal newline files, so that the diff does not go
sideways on us.

In Python 3.3 this was an option that you can pass to the ZipFile.open
() method to turn this on.

In Python 3.8 this is no longer directly supported and instead you get
this feature by using TextIOWrapper to read the file. We were already
doing that, and this class uses universal newlines by default, so this
is a small fix.