format() and some string concat become fstrings
Since we are now using a newer version of Python, the first round of
simple fixes involves replacing all of the obvious places where we were
using format() calls and string concatenation in places where we can
now use an fstring to get the same result but with cleaner code.

These changes were arrived at by doing a search for common constructs;
more likely exist in other places, and will be fixed as we run into
them going forward.