Fix issues with inconsistent hover popup data
As previously implemented, the hover popup on a package name could
either have detailed information in it (such as in a Diff report) or
simple information (such as the Package Report), where the difference
is whether or not a deep scan is done to count the exact number and
type of all overrides.

Package Reports would use a simple version since they display every
known package, which would be costly for little gain.

A downside to this is that the naive version passes an override count of
1 as a placeholder, which was making the popup display confusing
results (like telling you there are overrides when there are not).

Presumably this was a holdover from a previous version of the popup.

In any case, now the status() method returns -1 in this case rather than
one, includes information on whether overrides are possible, and the
hover popup uses a combination of both to know if it should display
details, or just a message on the possibility of an override.