Always display the Python line in the hover popup
In order to increase clarity (and decrease confusion), the hover popup
has been altered so that instead of hiding the Python line that says
what version of the plugin host the package runs in when there are no
plugins, that line appears but explicitly states that there are no
plugins.

This makes it clear what is happening, since otherwise it can be
confusing when you don't see the line.