Dynamically determine available Python hosts
Rather than workign under the assumption that packages are either 3.3 or
3.8, this code tries to determine, in as dynamic and future proof a
manner as is possible, what versions of Python exist as plugin hosts
so that appropriate versions can be selected.

The assumption is that if there are multiple versions (as there are
right now), only packages that declare explicitly get the higher
version and everything else gets the lower.

This will carry forward into the future, but the presumption is that
there will be only one host, and so everyone will use that.

The only outlier is that the .python-version file may cause problems
because currently Sublime blocks plugins from loading if it specifies
an invalid version of Python. If it does that going forward then all
3.8 packages will cease to work if their authors do not update them.