Clean up code to remove linter errors
Using the linter in LSP-pyright, all files have been cleaned as much as
I'm willing to clean to remove the number of warnings and errors that
they pop up.

During the course of this, a couple of actual legitimate edge case bugs
were found and fixed.

A few warnings still remain, some because Pyright is dumb, and some
because I refuse to fix them on general principles (but they are
harmless anyway).