Make clear what the dependency count is counting
The Package Report shows a list of installed dependencies, except that
if you are using Package Control 4 or higher, dependencies are usually
installed as actual libraries outside of the Packages folder, in which
case they don't show up in the list.

As such, the count was technically wrong since it was describing the
wrong thing; so the text has been updated.