Update the reports that exclude unchanged
This is a slight modification/bugfix to the bulk diff report that can
exclude files that are unchanged, and the override report that excludes
unchanged files.

In the first case we update the message on what's excluded that gets
displayed in the console so that it is more informative about what got
excluded.

In the second case, we made the same change but also fixes a bug that
was previously masked by Python 3.3 where we were altering the contents
of a PackageFileSet while we were iterating it, which causes a problem.