Shift location of our custom CSS
Locally, having the custom stylesheet that we use in the root of the
docs/ folder and not in content/ works, but when deployed out to
Cloudflare, that file is missing.

This is confusing because locally it seems present at the expected URL,
and in fact various other files get bundled that probably should not,
but that does not happen remotely, so it would appear. Or, we don't get
to see because it only uploads what it thinks it needs to, or such.

In any case, checking to see if this resolves the issue.