Change the documented minimum build number
The minimum build is being switched from 4126 to 4200. The original
intent of making 4126 the minimum was because new development was going
to start and we wanted to start with a known good version as a base.

Since we're spinning up version 3.0.0 to get ready for Sublime switching
to having only a single plugin host, we might as well get the version
number closer as well.

Thus far we have not done anything that would need the newer version, so
we are also open to being able to change it one more time once the
official plugin host drops, in theory.