Documentation Cleanup Pass
The prior commits put the bulk of the work in to bringing the original
documentation into this new documentation set.

This commit is a last pass prior to release to clean up as much as we
can (though some other additions may be pending) to bring the
documentation to a releasable state.

This includes cleaning up some text, adding in some new links, and
refreshing the screenshots that are in use to be more recent.

In theory, this gets us ready to release by just changing the
appropriate changelog entry, at this point.