Release version 3.0.0
This release fixes a smattering of bugs, cleans up the code and moves
the package into the 3.8 plugin host from the 3.3 plugin host to get
ready for the sunsetting of the 3.3 plugin host.

The code was updated with the intention that when 3.8 is dropped in the
next dev cycle, we will cleanly run in the new version.

As a part of this release we're changing the version bump that was added
back in 2.2.5 from 4126 to 4200.