Experiment with using mkdocs-macros plugin
Some parts of the documentation are a real pain in the butt to generate
links to, because all links need to be relative.

There is a plugin for mkdocs that allows you to include custom macros
that can be used to inject content in the page before it is converted.

This commit includes the plugin, and a sample module with a setting()
function macro that replaces the setting name you gave it with a link
to the page for that setting.

All of the documentation pages with settings mentioned have been
adjusted to use this.

This makes the link generation less error prone and also easier to read
in the source documentation files, which seems like a win.