Include a custom variable for the year
In order to make sure that the copyright on the license page in the
documentation is always up to date, this includes a new custom variable
that returns the current year when the documentation is generated.

The macro package can't touch the footer text because that's part of the
layout template and not part of the content of the page. So for that we
use a mkdocs configuration hook to dynamically update the footer config
from the configuration file and expand the year ourselves.