Extend the macro system to make command links easy
Similar to the previous change for easily linking to a setting, this
change allows for easily linking to a command based on its name in the
documentation.

This is done in such a way as to make the macro take the name of the
command and have it infer its own slug. The link will use the provided
command name as the link text unless you pass specific link text.

For orthogonality, the macro for settings is adjusted to work the same
way, even though currently that is not required.